distrib_systems_tasks/weeks/week-06/app/requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
httpx
pytest